Value_type_and_Reference_type_WorkingWithTypes fix: fixing errors in writing code

+++ Models/Curreny.cs
@@ -8,7 +8,7 @@ namespace Models
 {
     public struct Curreny
     {
-        public decimal Code { get; set; }
+        public int Code { get; set; }
         public string Name { get; set; }
 
     }
